requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
daphne==4.1.2
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
kafka-python==1.4.7
psycopg2-binary==2.9.10
redis==5.2.1
websockets==14.2